# docs/ai/contracts/praxis-ai-api-contract-v1.1.openapi.yaml
openapi: 3.0.3
info:
  title: Praxis AI Patch & Stream Contract
  version: v1.1
  description: |
    Fonte unica de contrato para endpoints corporativos de patch e streaming SSE.
    Cobre /patch retrocompativel, /patch/stream/{start,connect,probe,cancel}
    e /authoring/turn/stream/{start,connect,probe,cancel}.
servers:
  - url: /
...
paths:
  /api/praxis/config/ai/patch:
    post:
      tags: [ai-patch]
      operationId: generatePatch
      summary: Gera patch snapshot (retrocompativel)
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/EnvHeader'
        - $ref: '#/components/parameters/ContractVersionHeader'
        - $ref: '#/components/parameters/ContractSchemaHashHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiOrchestratorRequest'
      responses:
        '200':
          description: Patch/clarification/error gerado
          headers:
            X-Praxis-Contract-Version:
              $ref: '#/components/headers/ContractVersionHeader'
            X-Praxis-Schema-Hash:
              $ref: '#/components/headers/ContractSchemaHashHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiOrchestratorResponse'
        '409':
          description: Mismatch de contrato (versao/hash)
          headers:
            X-Praxis-Contract-Version:
              $ref: '#/components/headers/ContractVersionHeader'
            X-Praxis-Schema-Hash:
              $ref: '#/components/headers/ContractSchemaHashHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiOrchestratorResponse'
        '422':
          description: Erro de validacao semantica (ex. enum invalido)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiOrchestratorResponse'
...
  /api/praxis/config/ai/authoring/manifests/{componentId}:
    get:
      tags: [ai-authoring]
      operationId: getAgenticAuthoringManifest
      summary: Le o authoringManifest executavel projetado no ai_registry
      parameters:
        - $ref: '#/components/parameters/ComponentIdPath'
      responses:
        '200':
          description: Manifest executavel do componente
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Manifest inexistente ou invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringDryRunErrorResponse'
  /api/praxis/config/ai/authoring/manifests/{componentId}/editable-targets:
    get:
      tags: [ai-authoring]
      operationId: listAgenticAuthoringManifestTargets
      summary: Lista editableTargets declarados pelo manifest do componente
      parameters:
        - $ref: '#/components/parameters/ComponentIdPath'
      responses:
        '200':
          description: Alvos editaveis declarados pelo manifest
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '400':
          description: Manifest inexistente ou invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringDryRunErrorResponse'
  /api/praxis/config/ai/authoring/manifests/{componentId}/operations:
    get:
      tags: [ai-authoring]
      operationId: listAgenticAuthoringManifestOperations
      summary: Lista operacoes tipadas declaradas pelo manifest do componente
      parameters:
        - $ref: '#/components/parameters/ComponentIdPath'
      responses:
        '200':
          description: Operacoes tipadas declaradas pelo manifest
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '400':
          description: Manifest inexistente ou invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringDryRunErrorResponse'
  /api/praxis/config/ai/authoring/manifests/{componentId}/presentation-affordances:
    get:
      tags: [ai-authoring]
      operationId: listAgenticAuthoringManifestPresentationAffordances
      summary: Lista affordances de apresentacao governadas para o componente
      parameters:
        - $ref: '#/components/parameters/ComponentIdPath'
      responses:
        '200':
          description: Catalogo de affordances de apresentacao do componente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringPresentationAffordanceCatalog'
        '400':
          description: Catalogo inexistente ou invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringDryRunErrorResponse'
  /api/praxis/config/ai/authoring/manifests/{componentId}/resolve-target:
    post:
      tags: [ai-authoring]
      operationId: resolveAgenticAuthoringManifestTarget
      summary: Resolve alvo usando target resolver declarado no manifest
      parameters:
        - $ref: '#/components/parameters/ComponentIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgenticAuthoringResolveTargetRequest'
      responses:
        '200':
          description: Resultado de resolucao do alvo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringResolvedTarget'
        '400':
          description: Manifest inexistente, invalido ou requisicao inconsistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringDryRunErrorResponse'
  /api/praxis/config/ai/authoring/manifests/{componentId}/validate-plan:
    post:
      tags: [ai-authoring]
      operationId: validateAgenticAuthoringManifestPlan
      summary: Valida plano de edicao contra estrutura executavel do manifest
      parameters:
        - $ref: '#/components/parameters/ComponentIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgenticAuthoringManifestEditPlanRequest'
      responses:
        '200':
          description: Resultado de validacao estrutural do plano
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringManifestValidationResult'
        '400':
          description: Manifest inexistente, invalido ou requisicao inconsistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringDryRunErrorResponse'
  /api/praxis/config/ai/authoring/manifests/{componentId}/compile-patch:
    post:
      tags: [ai-authoring]
      operationId: compileAgenticAuthoringManifestPatch
      summary: Compila efeitos genericos do manifest em patch aplicavel
      parameters:
        - $ref: '#/components/parameters/ComponentIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgenticAuthoringManifestEditPlanRequest'
      responses:
        '200':
          description: Patch compilado com proposedConfig ou falhas de compilacao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringManifestCompileResult'
        '400':
          description: Manifest inexistente, invalido ou requisicao inconsistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgenticAuthoringDryRunErrorResponse'
...
components:
  parameters:
...
    ComponentIdPath:
      name: componentId
      in: path
      required: true
      schema:
        type: string
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: false
      schema:
        type: string
    UserIdHeader:
      name: X-User-ID
      in: header
      required: false
      schema:
        type: string
    EnvHeader:
      name: X-Env
      in: header
      required: false
      schema:
        type: string
    ContractVersionHeader:
      name: X-Praxis-Contract-Version
      in: header
      required: false
      schema:
        type: string
        default: v1.1
    ContractSchemaHashHeader:
      name: X-Praxis-Schema-Hash
      in: header
      required: false
      schema:
        type: string
        default: 41767dafdf6ebea802c083d77187ad6612937d6831188938dc21adce384b57a2
...
  headers:
    ContractVersionHeader:
      description: Versao de contrato efetiva no backend
      schema:
        type: string
        example: v1.1
    ContractSchemaHashHeader:
      description: Hash do schema de contrato efetivo no backend
      schema:
        type: string
        example: 41767dafdf6ebea802c083d77187ad6612937d6831188938dc21adce384b57a2
  schemas:
    AgenticAuthoringDryRunErrorResponse:
      type: object
      properties:
        valid:
          type: boolean
        failureCodes:
          type: array
          items:
            type: string
        message:
          type: string
      additionalProperties: false
    AgenticAuthoringManifestEditPlanRequest:
      type: object
      properties:
        config:
          type: object
          nullable: true
          additionalProperties: true
        plan:
          type: object
          nullable: true
          additionalProperties: true
        validationContext:
          type: object
          nullable: true
          additionalProperties: true
          description: >
            Contexto governado usado somente por validadores/compiladores de authoring.
            Não deve ser persistido no config materializado. Para praxis-table, pode
            carregar o schema request de filtro canônico em filterRequestSchema ou
            metadados equivalentes em filterRequestFieldMetadata, permitindo validar
            campos de filtro que não são colunas visíveis da tabela.
      additionalProperties: false
    AgenticAuthoringPresentationAffordanceCatalog:
      type: object
      properties:
        version:
          type: string
        kind:
          type: string
          enum: [praxis.ai-authoring.presentation-affordance-catalog]
        componentId:
          type: string
        defaultTargetKind:
          type: string
        sourceRef:
          type: string
        affordances:
          type: array
          items:
            $ref: '#/components/schemas/AgenticAuthoringPresentationAffordance'
      required:
        - version
        - kind
        - componentId
        - defaultTargetKind
        - sourceRef
        - affordances
      additionalProperties: false
...
    AgenticAuthoringResolveTargetRequest:
      type: object
      properties:
        config:
          type: object
          nullable: true
          additionalProperties: true
        operationId:
          type: string
        target:
          nullable: true
        input:
          nullable: true
      additionalProperties: false
    AgenticAuthoringResolvedTarget:
      type: object
      properties:
        status:
          type: string
        componentId:
          type: string
        operationId:
          type: string
        kind:
          type: string
        resolver:
          type: string
        path:
          type: string
        value:
          nullable: true
        candidates:
          type: array
          items:
            type: string
        failures:
          type: array
          items:
            type: string
      additionalProperties: false
    AgenticAuthoringManifestValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        failures:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        normalizedPlan:
          type: object
          additionalProperties: true
      additionalProperties: false
    AgenticAuthoringManifestCompileResult:
      type: object
      properties:
        compiled:
          type: boolean
        failures:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        patch:
          $ref: '#/components/schemas/AgenticAuthoringCompiledComponentPatch'
      additionalProperties: false
...
    AiOrchestratorRequest:
      type: object
      required:
        - componentId
        - componentType
        - currentState
      properties:
        componentId:
          type: string
        componentType:
          type: string
        userPrompt:
          type: string
        sessionId:
          type: string
          format: uuid
        mode:
          type: string
          enum: [new, continue]
        clientTurnId:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: '#/components/schemas/AiChatMessage'
        summary:
          type: string
        uiContextRef:
          $ref: '#/components/schemas/AiUiContextRef'
        currentStateDigest:
          $ref: '#/components/schemas/AiCurrentStateDigest'
        currentState:
          type: object
          additionalProperties: true
        dataProfile:
          type: object
          nullable: true
          additionalProperties: true
        schemaFields:
          type: array
          nullable: true
          items:
            type: object
            additionalProperties: true
        runtimeState:
          type: object
          nullable: true
          additionalProperties: true
        suggestedPatch:
          type: object
          nullable: true
          additionalProperties: true
        contextHints:
          type: object
          nullable: true
          properties:
            domainCatalog:
              $ref: '#/components/schemas/AiDomainCatalogContextHint'
          additionalProperties: true
        aiMode:
          type: string
        requireSchema:
          type: boolean
        resourcePath:
          type: string
        contractVersion:
          type: string
          default: v1.1
        schemaHash:
          type: string
          default: 2853423fce747f620266750bcf7f1387278afc7c94dee3a33d96da1a3f40f511
        schemaContext:
          $ref: '#/components/schemas/AiSchemaContext'
        variantId:
          type: string
        apiMethod:
          type: string
        apiTags:
          type: string
        apiSearchLimit:
          type: integer
          format: int32
      additionalProperties: false
    AiOrchestratorResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
          nullable: true
          description: >
            Codigo funcional estavel. A policy canonica de admissao pode retornar
            AI_TURN_POLICY_REJECTED antes de chamar provider LLM quando o pedido
            solicita segredos, prompt interno, bypass, logging oculto ou execucao
            destrutiva externa, e AI_TURN_OUT_OF_SCOPE quando o pedido publico nao
            puder ser aterrado no componente, contrato, manifest ou capabilities atuais.
        type:
          type: string
          enum: [patch, clarification, error, info]
        contractVersion:
          type: string
          nullable: true
        schemaHash:
          type: string
          nullable: true
        patch:
          type: object
          nullable: true
          additionalProperties: true
        componentEditPlan:
          type: object
          nullable: true
          additionalProperties: true
          description: Plano declarativo opcional retornado quando contextHints.authoringContract.preferredResponse=componentEditPlan. O adaptador do componente valida e compila o plano antes de aplicar.
        diff:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/AiPatchDiff'
        explanation:
          type: string
          nullable: true
        warnings:
          type: array
          nullable: true
          items:
            type: string
        message:
          type: string
          nullable: true
        options:
          type: array
          nullable: true
          items:
            type: string
        optionPayloads:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/AiOption'
        contextRequest:
          type: array
          nullable: true
          items:
            type: integer
        clarification:
          $ref: '#/components/schemas/AiClarificationUi'
        componentId:
          type: string
          nullable: true
        componentType:
          type: string
          nullable: true
        path:
          type: string
          nullable: true
        providedValue:
          nullable: true
        allowedValues:
          type: array
          nullable: true
          items:
            type: string
        memory:
          $ref: '#/components/schemas/AiMemoryInfo'
        observationId:
          type: string
          format: uuid
          nullable: true
          description: Identificador redigido de observabilidade para feedback/triagem do assistente.
      additionalProperties: false
...
    AiPatchDiff:
      type: object
      required: [path]
      properties:
        path:
          type: string
        before:
          nullable: true
        after:
          nullable: true
      additionalProperties: false
    AiOption:
      type: object
      properties:
        value:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        example:
          type: string
          nullable: true
        contextHints:
          type: object
          nullable: true
          additionalProperties: true
      additionalProperties: false
    AiClarificationUi:
      type: object
      properties:
        responseType:
          type: string
          enum: [text, choice, confirm, mixed, context]
        selectionMode:
          type: string
          enum: [single, multiple]
        presentation:
          type: string
          enum: [buttons, list, chips]
        allowCustom:
          type: boolean
          nullable: true
      additionalProperties: false
...
    AiSchemaContext:
      type: object
      properties:
        path:
          type: string
          nullable: true
        operation:
          type: string
          nullable: true
        schemaType:
          type: string
          nullable: true
      additionalProperties: false
    AiChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
      additionalProperties: false
    AiUiContextRef:
      type: object
      properties:
        componentType:
          type: string
          nullable: true
        componentId:
          type: string
          nullable: true
        routeKey:
          type: string
          nullable: true
        schemaHash:
          type: string
          nullable: true
        variantId:
          type: string
          nullable: true
      additionalProperties: false
    AiCurrentStateDigest:
      type: object
      properties:
        columns:
          type: array
          nullable: true
          items:
            type: string
        sort:
          type: string
          nullable: true
        rowCount:
          type: integer
          nullable: true
      additionalProperties: false
...